enhanced email sending features.
