+++ server/Auth/register.py
@@ -1,13 +1,14 @@
 from flask_restful import Api, Resource
 from flask import make_response, request
 from http import HTTPStatus
-from .emailbot import EmailBot
+from ..EmailBot.emailbot import EmailBot
 
 
 class RegisterRoute(Resource):
     def post(self):
-        print(request.values)
-        # result = database.users.find_one({'username': u})
-        # with EmailBot() as email:
-            # email.send_register()
+        print(request.json)
+        username = request.json['username']
+        # result = database.users.find_one({'username': u})  <--- Need DB setup
+        with EmailBot(purpose='registration') as email:
+            email.send(username)
         return make_response("Post works\n", HTTPStatus.OK)
