+++ server/EmailBot/registration.html
@@ -1,0 +1,9 @@
+<html>
+  <body>
+    <p>Hi,<br>
+       This is the registration page for Deerhunt<br>
+       <a href="https://utmrobotics.com/">UTM Robotics</a>
+    </p>
+    <img src="https://camo.githubusercontent.com/00ece4e61c53cd5e25c8dc5ad997e1024b5255a71ea86f2e180a8b06b4c19858/68747470733a2f2f7468756d62732e647265616d7374696d652e636f6d2f622f636f6c6f7266756c2d666162756c6f75732d776f72642d3138303239333236352e6a7067">
+  </body>
+</html>
